pyproject.toml
[tool.poetry]
name = "ChatIQ"
version = "0.1.0"
description = "A versatile Slack bot using GPT & Weaviate-powered long-term memory to accomplish various tasks."
authors = ["yujiosaka <yujisobe@gmail.com>"]
license = "GPL-3.0"
readme = "README.md"
homepage = "https://github.com/yujiosaka/ChatIQ"
repository = "https://github.com/yujiosaka/ChatIQ"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.10",
]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
cryptography = "^40.0.2"
flask = "^2.3.2"
gunicorn = "^20.1.0"
langchain = "^0.0.173"
openai = "^0.27.6"
psycopg2-binary = "^2.9.6"
python-dotenv = "^1.0.0"
requests = "2.28.0" # Pin to 2.28.0 due to a bug introduced in 2.30.0 (https://github.com/psf/requests/issues/6437)
slack-bolt = "^1.18.0"
slack-sdk = "^3.21.3"
sqlalchemy = "^2.0.13"
tiktoken = "^0.4.0"
pymupdf = "^1.22.3"
python-dateutil = "^2.8.2"
pytz = "^2023.3"
weaviate-client = "^3.17.1"
beautifulsoup4 = "^4.12.2"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
coverage = "^7.2.5"
flake8 = "^6.0.0"
freezegun = "^1.2.2"
isort = "^5.12.0"
pre-commit = "^3.3.1"
pytype = "^2023.5.8"
pytest = "^7.3.1"
pytest-mock = "^3.10.0"
requests-mock = "^1.10.0"

[tool.poetry.scripts]
chatiq = "bin.cli:run"

[tool.black]
line-length = 125

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
